Add: GetSearchQuery function
Add : GetSearchQuery function.

that allows for  getting list of information in a collection of files that fulfil some query specifications.

Add: OptionsAdd, SelectAdd
Commands that behave like Select and Options.

Adds the ability for the user to add more option/s to a list if they do are not found.

Change: When displaying options, and Select prompts, now it uses the suggester, instead of asking for the number of an option.

This should help with big option lists, and reduce the dependency on the user to type the right option number.